refactor(realestate components): folder restructuring
